fix: allow decimal amounts in quote form and money columns
The operator quote form rejected decimal amounts (e.g. ₦1,250.50) at
three layers:

1. Database: all NGN money columns were integer type, which silently
   truncates or rejects decimals. Migration 0048 changes them to
   numeric(12,2) across assisted_order_lines, assisted_order_quotes,
   assisted_order_line_verifications, assisted_order_payments, and
   service_fee_policies.

2. Zod schema: moneyComponent used .int() which rejected decimals
   before they reached the database. Removed .int() from
   moneyComponent and serviceFeePolicyResolvedNgn. Also removed .int()
   from the service fee policy action schema.

3. Form: inputMode='numeric' showed a phone keypad without a decimal
   key, and Number.isInteger() validation rejected any decimal input.
   Changed to inputMode='decimal', added step='0.01', and changed
   validation to Number.isFinite(). Applied the same fixes to the
   service fee manager form.

Also fixed the Paystack kobo conversion to use Math.round() when
converting naira to kobo, preventing floating-point drift on decimal
amounts.

+++ app/(ops)/ops/service-fees/actions.ts
@@ -20,10 +20,10 @@ const policySchema = z.object({
   retailerSlug: z.string().trim().max(120).nullable().optional(),
   deliveryState: z.string().trim().max(60).nullable().optional(),
   feeModel: feeModelEnum,
-  flatFeeNgn: z.number().int().min(0).nullable().optional(),
+  flatFeeNgn: z.number().min(0).nullable().optional(),
   percentageRate: z.number().min(0).max(100).nullable().optional(),
-  minFeeNgn: z.number().int().min(0).nullable().optional(),
-  maxFeeNgn: z.number().int().min(0).nullable().optional(),
+  minFeeNgn: z.number().min(0).nullable().optional(),
+  maxFeeNgn: z.number().min(0).nullable().optional(),
   priority: z.number().int().min(0).max(1000).default(0),
   notes: z.string().trim().max(1000).nullable().optional(),
 });
